Change the index to d from 0 to ctr.
WIth changing the index to the dictionary, the program used up a lot of CPU time doing nothing.
Added a status line every once in a while.  Added formatting for the status line.

CAUTION: This program may use up all of the RAM and swap space on your machine, creating an Out-of-Memory condition.  If this happens, then your machine may lockup or freeze, but you might see a great deal of disk activity.